+++ NEWS.md
@@ -1,3 +1,7 @@
+# fluxible 1.3.2
+
+* `flux_plot`: fixed a bug, which makes it now much faster
+
 # fluxible 1.3.1
 
 * Vignettes and readme: reference and small fixes